pod_automation_api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-jose>=3.3.0
python-multipart>=0.0.6
httpx>=0.24.0
pyyaml>=6.0
email-validator>=2.0.0
